tesseract_ocr/requirements.txt
Flask>=2.3.2
numpy>=1.22.0
pillow>=6.2.0
pytesseract>=0.2.6
